backend/requirements.txt
# CORE OPC Calculator Backend Dependencies
# Exact versions as specified in cursorrules.md

# FastAPI and ASGI
fastapi==0.110.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.36
psycopg[binary]==3.2.10
alembic==1.13.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# HTTP Client
httpx==0.27.0

# Environment Variables
python-dotenv==1.0.1

# CORS
python-multipart==0.0.9

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.0

# Development
python-multipart==0.0.9
